Move expand_dag_tags to common.py and tests to test_task_runner.py
