--- shared/observability/src/airflow_shared/observability/attributes.py
@@ -1,28 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-from __future__ import annotations
-
-from collections.abc import Iterable
-
-
-def expand_dag_tags(tag_names: Iterable[str]) -> dict[str, str]:
-    """Expand DAG tag attributes into key-value pairs."""
-    result: dict[str, str] = {}
-    for name in tag_names:
-        key, _, value = name.partition(":")
-        result[key] = value
-    return result
